Make sure hypervisor is excluded from ocount and operf
Since we have no interface support in the event specification to
allow the user to select or de-select counting events in hypervisor,
and also since the output of ocount and opreport do not support the
concept of hypervisor, we should exclude hypervisor from counting
and profiling.  There's a bug in the current code such that the
user may or may not get hypervisor events included.  This patch
explicitly excludes hypervisor.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>