Add oprofile support for Broadwell microarchitecture
This patch adds the event list of the Intel Broadwell architecture.
Hopefully this can still make 1.0

The patch is very straight forward: just add the model numbers and
type in the usual places and add the event list.

Passes make check

Some notes:
- Haswell included one Broadwell model number by mistake. I moved
that to Broadwell now.
- oprofile doesn't support umask sub events with different counter
constraints than other events. This affects a few events on Broadwell.
However it's not a problem when oprofile uses perf as a backend,
as perf will know how to schedule these events (once it gets the
Broadwell support). It won't work correctly with the old driver.
Most of these events are not too useful for sampling, so in practice
it's not a real problem.
- As usual PEBS events and events with offcore mask and uncore
events are missing.

Signed-off-by: Andi Kleen <ak@linux.intel.com>