Update the Haswell events to the latest version
Some minor changes to the previous version, but it should be more
consistent with other tools now.

The event name descriptions have been dropped. They were never all that
useful anyways because the event is defined by the unit masks.
Now all events with more than one unit mask only have a description
in the unit masks.

As a new feature any known Errata to the event are referenced.

Signed-off-by: Andi Kleen <ak@linux.intel.com>