operf main process improperly killing conversion process
Unless operf is run with --lazy-conversion, a separate process (the
'operf_read_pid' as referred to in the code) is forked to read
perf_events-formatted kernel data from a pipe.  The pipe writer is
another oprofile process  -- operf_record_pid -- which is reading the
actual kernel data.  The operf_read_pid converts this data to oprofile
sample files.  This conversion process is fairly expensive, so it's not
unusual for it to take a little while after profiling stops for it to
complete.  And the more samples there are (i.e., the longer the profile runs),
the longer while it will take.  When profiling stops, we call
_kill_operf_read_pid and a 5 second timeout begins, waiting for operf_read_pid
to finish its conversion. If it has not completed in 5 seconds, we kill the
operf_convert_pid!

Having a hard-coded timeout was a bad idea; we should always allow the
conversion process to finish. It should eventually read all the data from the
pipe and then detect that the write has closed the pipe.  At that point,
it will exit on its own.

This patch changes the name of the _kill_operf_read_pid() function to
_waitfor_operf_read_pid(), and modifies its behavior to simply wait for
the read process to finish and check its exit status.  The conversion
process has been slightly modified so that, when it's in post-profiling
conversion phase, it will print a "." every 1 million records processed
to show its progress.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>
Reported-by:   Andrew Haley <aph@redhat.com>