Remove hard-coded timeout for JIT dump conversion
A problem was reported by a user profiling a Java application where
not all JIT dump files were being converted to <PID>.jo files.
This problem is similar in nature to the problem fixed in an Aug 12
commit ("operf main process improperly killing conversion process"),
which involved an ill-advised timeout where the main operf thread
is waiting for the oper_read_pid to finish. Now, another unnecessary
timeout has been encounterd where the operf_read_pid is waiting for the
JIT dump conversion process to finish. We should always allow the JIT
dump conversion to finish. If it hangs, the user can always clt-c to
abort the profile run. A hang should never occur during JIT dump conversion,
so that would be a bug that the user should report. This patch removes
this timeout.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>