Make opagent error messages more helpful
While helping a user with a problem involving libopagent,
I noticed that opagent.c would print identical messages
at various places in the code, making it impossible to
determine exactly where the message was coming from.
This patch adds a "#<number>" tag to the messages so
they can be uniquely identified for debugging purposes.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>