Java profiling: opagent: fwrite_unlocked failed
Certain Java Virtual Machines do not provide full code information in
the JNI callbacks to oprofile's java agent library; in particular,
the code size argument may be set to zero even if the code address
is non-null. The libopagent/opagent.c:op_write_native_code function
was not properly handling this case and would print the message
   opagent: fwrite_unlocked failed
and return a -1, causing the caller to display a message like
   Error: op_write_native_code(): Success

This patch fixes that issue.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>