Fix configure script to error out when no perf_events support is found
Prior to release 1.0.0, if the configure script found that the kernel
version was not new enough to have perf_events support or if the perf_event.h
header file could not be found, we would just fall back to building only
the legacy opcontrol-based profiler and would skip building operf and ocount.
But as of release 1.0.0, the opcontrol-based profiler is no longer included,
so if operf and ocount cannot be built, we should error out of the configure
execution, which is what this patch does.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>