Flesh out oparchive procedure and fix other minor doc errors
The oparchive man page does not do a very good job of documenting
how oparchive is intended to be used.  This patch addresses that
issue, in particular by directing the reader to the user manual
where there is much more detail to be found.

Additionally, this patch makes a few other minor doc fixups,
such as fixing an example shell script in the chapter on opimport
(a comment line did not have a "#" at the beginning of the line)
and removing a reference to qt stuff in the user manual. The patch
also cleans up the oprofile man page, updating it to reflect the
change from opcontrol operf.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>