Add a check to ensure event codes in events files fit in an integer
The read_events() function in libop/op_events.c is enhanced to
validate that the hex codes found in events files are not too big
to fit in an integer, since the rest of oprofile assumes hex codes
are integers.  This check will execute whenever ocount or operf is
run, as well as when doing 'make check' or 'make distcheck'.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>