Passing a bogus value to configure --with-java should cause  error
The "--with-java" configure option expects either "no", "yes", or a
path to a directory which has an 'include' directory containing either
jvmpi.h or jvmti.h or both. If the user who is running 'configure'
specifies an incorrect pathname, the configure script continues on to
the end and exits normally with no error message. If the user looks
carefully at the output of the configure script, they will see messages
indicating jvmpi.h and jvmti.h were not found, but that's not sufficient
warning to the user that something went wrong.

This patch will detect an invalid Java home pathname and exit immediately
with an error.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>