Avoid permanently setting POSIXLY_CORRECT environment variable
During testing on Fedora it was discovered that operf was setting the
enviroment variable POSIXLY_CORRECT and this could potentially be
observed in the children tasks that operf starts (Red Hat Bugzilla
1178577). The operf, ocount, and opjitconv commands all ensure that
POSIXLY_CORRECT environment variable is set when the options are
processed with getopt_long, but they never unset the variable
afterwards.  This patch ensures that POSIXLY_CORRECT is as it was
before it was set.

Signed-off-by: William Cohen <wcohen@redhat.com>