Correctly track the size of operf poll_data array
The new[] operator in C++ returns a pointer to a block of memory
allocated for an array of objects.  If the sizeof operator is used on
the value returned by new[], the sizeof operator will return the space
required for the pointer rather than the space required for the
array. Thus, the following code will not correctly count the number of
elements in the array:

	struct pollfd * poll_data;
	poll_data = new struct pollfd [100];
	num_mmaps = sizeof(poll_data)/sizeof(poll_data[0]);

To properly track the size of the dynamically allocated poll_data
array a num_mmaps field has been added to the operf_record class.

Signed-off-by: William Cohen <wcohen@redhat.com>