Fix default unit masks for Intel Broadwell
Since some of the default unit masks for Intel Broadwell events cannot be
uniquely specified by numbers, the defaults have had to be replaced
by the named ones. When the affected events are used on Broadwell without
specifying unit masks after applying this patch, the default masks
are chosen correctly.

Signed-off-by: William Cohen <wcohen@redhat.com>