Improve handling of remapped anon regions across processes
Java runtime environments use dynamically allocated memory in
anonymous regions to store Just-In-Time translated code.  The Java
runtime system may change the access permissions for portions of mmap
regions during execution and operf needs to be tolerant of those
change to a portion of the mmap.  operf also needs to keep the anon
memory maps distinct between processes to avoid confusion about the
sizes of the memory regions.

Signed-off-by: William Cohen <wcohen@redhat.com>