Fully initialialize the magic value for the operf_record header
On x86-64 machine operf would generate the following message and not
record any profiling data:

Error: operf sample data does not have expected header dataoperf-record process

The problem was tracked to the header magic value setting, which has a
string of 7 bytes rather than 8 bytes required for the u64 value it
was typecast to.  The code also made the assumption that processors
can do unaligned memory accesses when copying the string to the u64
magic field in the f_header struct.  This change ensures that the
value is fully initialized and that unaligned access support is not
required.

Signed-off-by: William Cohen <wcohen@redhat.com>