Avoid duplicate events for arm/armv7-ca7
For the armv7-ca7 processor the arm/armv7-common event file is include
twice.  One arm/armv7-common include is in the armv7-ca7 event file
itself and another include is in the included arm/armv7-ca5 event
file.  Removed the one in the arm/armv7-ca7 event file so the
armv7-common is only included once to eliminate the duplicate events.

Signed-off-by: William Cohen <wcohen@redhat.com>