Remove Powerpc OProfile events the kernel will reject
This patch comments out a number of events with "## note 1" to indicate
that the events are not supported.  A "note 1" is added to the beginning
of the file explaining the issue.  The issue is the events require the
setting of a register by the kernel when setting up to measure the
events that is only writable by the hypervisor.  Currently, there is no
API allowing the kernel to request the required field of the register be
changed by the hypervisor.  The events can be re-enabled if an API is
created for the OS to request the hypervisor set the bits.

Signed-off-by: Carl Love <carll@us.ibm.com>