Remove Powerpc OProfile events the kernel will reject
Will, Rei:

As I suspected in the last email, there is a second issue that has to do
with what version of the libpfm library OProfile is using.  Initially a
subset of the OProfile events for Power 8 were added to libpfm4.5.
Later the complete set of events was added to libpfm4.6.  So, My first
attempt at removing the events that perf was rejecting inadvertently
included events that were in libpfm 4.6 but not libpfm 4.5.  My version
of oprofile was used a patched version of libpfm 4.4 that effectively
made it libpfm 4.5.  I redid the patch an verified that when Oprofile is
built with libpfm 4.6 there are no rejected events.  I added a comment
in the event file as a heads up to this effect.

Please take a look at the patch and see if it works OK for you.  Sorry
for the delay in getting this out.

           Carl Love

----------------------------

Re-enable Power 8 events that the kernel does not reject.

The previous patch to remove Power 8 events that were being rejected by
the kernel also removed events that were actually being rejected by
OProfile.  OProfile was rejecting the events on the test machine because
the test machine used a version of libpfm that did not have all of the
Power 8 events.  This patch re-enables the Power 8 events that are
not rejected by the kernel.

Libpfm 4.5 only contains a subset of all the available Power 8 events.
The complete list of Power 8 events is supported by libpfm 4.6.  To use
all of the events in this file, OProfile must be compiled with
the libpfm 4.6 library or newer.  Otherwise, OProfile will reject the
event if it is not in the libpfm 4.5 library.

Signed-off-by: Carl Love <carll@us.ibm.com>