operf: prefer _stext over _text as the kernel address range start
In newer kernels, _stext is the right pseudo-symbol for marking the
start address of the kernel address range. The older _text symbol is
rather architecture specific.

In newer kernels, _stext is the right pseudo-symbol for marking the
start address of the kernel address range. The older _text symbol is
rather architecture specific.

With this patch, the _stext symbol is preferred (if found). In case
_stext is not there (an older kernel), _text is accepted like before.

With this patch, the _stext symbol is preferred (if found). In case
_stext is not there (an older kernel), _text is accepted like before.