Store samples in the archive and search the appropriate places for samples
Newer versions of oprofile use a oprofile_data directory in the
current working directory to store the samples.  This presents a
complications when data archived with oparchive.  The oparchive needs
to include samples in the archive.  The code also needs to make sure
that samples in the archive are used and not samples from a
oprofile_data directory in the current working directory.

Signed-off-by: William Cohen <wcohen@redhat.com>