Only start the application if the perf events setup was successful
The code was starting the application before the performance events
were setup.  In some cases the the setup of the perf events may fail
and the code needs to verify that the performance events have been
successfully set up before starting the application.  Changed the
order of those steps to allow a check of the perf event setup before
launching the application.

Signed-off-by: William Cohen <wcohen@redhat.com>