Fix OProfile libabi issues reported by GCC 7
GCC 7 uses the C++11 standards and dynamic exception specification are
deprecated.  Spelling out what exception the method throws would cause
GCC 7 to warn about this deprecated use and the make compiler setting
woudl cause the build to fail on the warning.

Also needed to explicitly pass in a couple parameters for the abi need
method as strings rather than char * pointers to get the code to
compile with GCC 7.

Signed-off-by: William Cohen <wcohen@redhat.com>