Correct a conditional check for ff_linenr_info
The GCC 7 compiler pointed out that the enum value ff_linenr_info was
being checked as a boolean by itself.  The ff_linenr_info enum is a
bit flag used to check whether a particular bit in the flags variable
is set.  This conditional was corrected to match other uses of
ff_linenr_info in format_output.cpp.

Signed-off-by: William Cohen <wcohen@redhat.com>