1/2 Oprofile support for Power9
Hi,

Add initial support for PPC64/Power9 to oprofile.
This includes the documentation updates to indicate "power9",
the PVR values, and all the other updates that will be necessary.
(This is based on a review of changes made when Power8 support
was initially added).
The long-ish event list follows as a subsequent patch.

Signed-off-by:  Will Schmidt <will_schmidt@vnet.ibm.com>

--