PowerPC: Remove trailing comma in cpu_name.
Will:

I redid the patch as follows per your suggestions.  I was able to get
on to a system where the issue occurs and test the patch.  The patch
seems to work fine.

                   Carl Love

-----------------------------------------------------------------------

The processor type for some Power 9 systems had a comma after POWER9.
Remove the comma before returning the string for the CPU.