Initialize the trans structure fields to values unused by perf samples
On newer 4.14 kernel operf was crashing because the kernel commit
1d953111b648e ("perf/core: Don't report zero PIDs for exiting tasks")
made perf samples have TID/PID of -1 for exiting processes.  OProfile
used this same value internally to indicate the that field had not
been initialized.  As a result operf would incorrectly attempt to run
some bookkeeping code due to an invalid match. The initialization of
the trans tgid field is now changed to a value that is unsed by the
perf samples to ensure that these invalid matches do not occur.

Signed-off-by: William Cohen <wcohen@redhat.com>