Handle empty event name spec gracefully for PPC
During testing Michael Petlan found that the giving an empty event
name in the event spec to operf on PPC machines would cause operf
crash.  Running the following following command would demonstrate
this problem:

  operf -e : /bin/ls

The problem was traced to event_name in _handle_powerpc_event_spec
function having a length of 0.  The strncmp using event_name in
_handle_powerpc_event_spec would indicate a match and cause strncpy to
be called to copy zero bytes. The zero length strncpy would crash
operf.  Added a check to make ensure that the string is non-zero
length.

Signed-off-by: William Cohen <wcohen@redhat.com>