Only start the application if the perf events setup was successful
Changes the order of starting the application and performance events.

Given this change we have a new issue.  The issue is the routine
start_counting() calls fork, creating app_PID process.

The parent then tries to setup the performance events, then if the
performance events were setup correctly, app_PID is then told to start
before exiting.  If the performance counter setup fails, the app_PID is
left running.  The app_PID is never told to start the workload, which is
correct but we don't record the fact that app_PID is running.  The
error path then fails to kill app_PID in routine main(), in file
oprofile-git/pe_counting, at about lie 909 because the if statement

 if (startApp && app_started && (run_result != APP_ABNORMAL_END)) {

is false because app_started is False.

The fix, I believe, is to set app_started to True in the parent code if
the fork was successful.  With this fix, there is no orphan processes
left after ocount exits.