Proposed fix for oprofile JIT bad event spec error
Hi Oprofile-devs,
Attached is a fix for a problem we have seen recently when
trying to profile against java/JIT workloads.

The error floated to the user at opreport time is:

  opreport error: parse_event_spec(): bad event specification: 165624.0x

Where the 165624 value will match the PIDs of the processes being profiled.

This ends up being an issue with how we are building the string for the file
that contains the ELF blob for the java/JIT code (i.e.  12345.jo).  We were
not truncating our 'path' string such that when we strncpy'd it into our
path-and-file variable, we were getting part of an address range (12345.0x)
instead of the desired .jo suffix.

The code has been updated to correct that.  Sniff testing shows we are
now successfully processing the .jo contents, and no longer seeing the bad
event spec error.

Tested locally, this appears to resolve the issue.

    agents/jvmti/libjvmti_oprofile.c:
     - update strncpy call to avoid gcc warning on strncpy parameters.
    opjitconv/opjitconv.c:
     - update logic around calculation of elf_file_size, and rework snprintf
     calls so we properly truncate the input string.

     Signed-off-by: Will Schmidt  <will_schmidt@vnet.ibm.com>