Use the mmap offset to correctly compute the IP location in a file
Newer versions of binutils are now now splitting the file into
multiple mmap loads.  The assumption that the mmap for the executable
code in the file starts at the beginning of the file (that the file
offset is always zero) is no longer true.  The computation to convert
the IP address into an offset needs to also use the offset.